backend/requirements.txt
numpy
pandas
matplotlib
requests
scikit-learn
# tensorflow
keras
scipy
flask
joblib
bcrypt
PyJWT
pymongo
flask_cors
python-dotenv
flask-limiter
gridfs
gunicorn
